django_project/requirements.txt
Django>=1.8,<2.0
djangorestframework
django_compressor
gunicorn